+++ auramoon/src/main/java/com/AuraMoon/auramoon/fnb/dto/MealOrderRequest.java
@@ -4,32 +4,32 @@ import java.util.List;
 
 public class MealOrderRequest {
 
-    private Long guestId;
-    private Long bookingId;
-    private List<Long> menuItemIds;
+    private Integer guestId;
+    private Integer bookingId;
+    private List<Integer> menuItemIds;
     private String note;
 
-    public Long getGuestId() {
+    public Integer getGuestId() {
         return guestId;
     }
 
-    public void setGuestId(Long guestId) {
+    public void setGuestId(Integer guestId) {
         this.guestId = guestId;
     }
 
-    public Long getBookingId() {
+    public Integer getBookingId() {
         return bookingId;
     }
 
-    public void setBookingId(Long bookingId) {
+    public void setBookingId(Integer bookingId) {
         this.bookingId = bookingId;
     }
 
-    public List<Long> getMenuItemIds() {
+    public List<Integer> getMenuItemIds() {
         return menuItemIds;
     }
 
-    public void setMenuItemIds(List<Long> menuItemIds) {
+    public void setMenuItemIds(List<Integer> menuItemIds) {
         this.menuItemIds = menuItemIds;
     }
