+++ auramoon/src/main/java/com/AuraMoon/auramoon/fnb/dto/MealSelectionRequest.java
@@ -5,26 +5,26 @@ import java.util.List;
 
 public class MealSelectionRequest {
 
-    private Long guestId;
-    private Long bookingId;
+    private Integer guestId;
+    private Integer bookingId;
     private LocalDate mealDate;
     private String mealType;
-    private List<Long> menuItemIds;
+    private List<Integer> menuItemIds;
     private String note;
 
-    public Long getGuestId() {
+    public Integer getGuestId() {
         return guestId;
     }
 
-    public void setGuestId(Long guestId) {
+    public void setGuestId(Integer guestId) {
         this.guestId = guestId;
     }
 
-    public Long getBookingId() {
+    public Integer getBookingId() {
         return bookingId;
     }
 
-    public void setBookingId(Long bookingId) {
+    public void setBookingId(Integer bookingId) {
         this.bookingId = bookingId;
     }
 
@@ -44,11 +44,11 @@ public class MealSelectionRequest {
         this.mealType = mealType;
     }
 
-    public List<Long> getMenuItemIds() {
+    public List<Integer> getMenuItemIds() {
         return menuItemIds;
     }
 
-    public void setMenuItemIds(List<Long> menuItemIds) {
+    public void setMenuItemIds(List<Integer> menuItemIds) {
         this.menuItemIds = menuItemIds;
     }
