+++ auramoon/src/main/java/com/AuraMoon/auramoon/fnb/repository/MealOrderRepository.java
@@ -3,5 +3,5 @@ package com.AuraMoon.auramoon.fnb.repository;
 import com.AuraMoon.auramoon.fnb.entity.MealOrder;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MealOrderRepository extends JpaRepository<MealOrder, Long> {
+public interface MealOrderRepository extends JpaRepository<MealOrder, Integer> {
 }
