+++ auramoon/src/main/java/com/AuraMoon/auramoon/fnb/repository/MenuItemRepository.java
@@ -3,5 +3,5 @@ package com.AuraMoon.auramoon.fnb.repository;
 import com.AuraMoon.auramoon.fnb.entity.MenuItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MenuItemRepository extends JpaRepository<MenuItem, Long> {
+public interface MenuItemRepository extends JpaRepository<MenuItem, Integer> {
 }
